CLICKHOUSE-3878: Move connection string validation to common, remove redundant headers, fix compilation issues

+++ dbms/programs/odbc-bridge/ODBCBridge.h
@@ -28,7 +28,7 @@ private:
 
     bool is_help;
     std::string hostname;
-    UInt16 port;
+    size_t port;
     size_t http_timeout;
     std::string log_level;
     size_t max_server_connections;
